+++ src/routes/api/v1/capabilities.rs
@@ -4,12 +4,12 @@
 
 use crate::database::models::prelude::*;
 use crate::database::DBConnection;
-use crate::http::responders::api::ApiResponse;
 
 use crate::guards::auth::Auth;
 
-use rocket::http::Status;
+use crate::http::responders::ApiResult;
 use rocket::Route;
+use rocket_contrib::json::Json;
 
 /// Collect every routes that this module needs to share with the application
 /// The name `collect` is a project convention
@@ -22,6 +22,11 @@ pub fn collect() -> Vec<Route> {
 /// The user needs to be authenticated, but this call do not require a special capability
 /// All the capability stored in database are responded into the json format
 #[get("/api/v1/capability")]
-pub fn get(conn: DBConnection, _auth: Auth) -> ApiResponse {
-    ApiResponse::new(Status::Ok, json!(CapabilityEntity::all(&conn).unwrap()))
+pub fn get(conn: DBConnection, _auth: Auth) -> ApiResult<Vec<Capability>> {
+    Ok(Json(
+        CapabilityEntity::all(&*conn)?
+            .into_iter()
+            .map(move |capability_entity| Capability::from(capability_entity))
+            .collect::<Vec<Capability>>(),
+    ))
 }